adicionar sistema de modificar e manter o modificador de preco por veiculo

+++ src/main/java/veiculo/Ticket.java
@@ -17,7 +17,7 @@ public class Ticket {
     }
 
     public double calcularValor() {
-        float modificador = vaga.getOcupante().getTipoDeVeiculo().getModificador();
+        double modificador = vaga.getOcupante().getTipoDeVeiculo().getModificador();
         float horas = ChronoUnit.HOURS.between(entrada, saida);
         if (horas == 0) {
             horas++;
